refactor(components): 将符号相关组件移动到 symbols 子目录
