+++ src/components/symbols/SymbolCard.tsx
