+++ src/components/symbols/SymbolDetail.tsx
