+++ src/components/symbols/SymbolList.tsx
@@ -2,8 +2,8 @@
 
 import { useState } from 'react';
 import { SymbolData } from '@/lib/core/types';
-import SymbolCard from '@/components/SymbolCard';
-import SymbolDetail from '@/components/SymbolDetail';
+import SymbolCard from './SymbolCard';
+import SymbolDetail from './SymbolDetail';
 
 interface SymbolListProps {
   displayedSymbols: SymbolData[];
