+++ src/components/symbols/index.ts
@@ -1,0 +1,3 @@
+export { default as SymbolList } from './SymbolList';
+export { default as SymbolCard } from './SymbolCard';
+export { default as SymbolDetail } from './SymbolDetail';
